Install patched Nerd-fonts in case when they are not installed

+++ Desktop_1/Install/1_install.sh
@@ -47,14 +47,17 @@ echo "Setup Termite"
 mkdir $XDG_CONFIG_HOME/termite
 yes | cp -arf "$termite_path""/." $XDG_CONFIG_HOME/termite
 
-echo "Setup patched Nerd-fonts. Can take long time..."
-cd ~/Downloads
-git clone https://github.com/ryanoasis/nerd-fonts --depth 1
-cd nerd-fonts
-./install.sh
-cd ..
-rm -rf nerd-fonts
-sudo mv ~/.local/share/fonts/* /usr/share/fonts
+if [[ ! -d "/usr/share/fonts/NerdFonts" ]]
+then
+    echo "Setup patched Nerd-fonts. Can take long time..."
+    cd ~/Downloads
+    git clone https://github.com/ryanoasis/nerd-fonts --depth 1
+    cd nerd-fonts
+    ./install.sh
+    cd ..
+    rm -rf nerd-fonts
+    sudo mv ~/.local/share/fonts/* /usr/share/fonts
+fi
 
 echo "Setup Neovim"
 mkdir -p $XDG_DATA_HOME/nvim/plugged
